修改头像显示问题，修改了一些bug

+++ zeus-server/src/main/java/com/nzl/server/service/impl/FileServiceImpl.java
@@ -62,6 +62,10 @@ public class FileServiceImpl implements FileService {
             return -1;
         }
 
+        if (filePath.contains("head")) {
+            return 1;
+        }
+
         // 文件存在则删除
         if (fileDir.delete()) {
             // 删除成功 返回 1
